fix: change Navlinks text, move Statistic to Profile
