--- src/pages/Dictionary/Statistics.tsx
@@ -1,7 +1,0 @@
-const Statistics = (): JSX.Element => (
-  <section className='difficult-words w-full flex flex-col items-center'>
-    <h1>Stats</h1>
-
-  </section>
-);
-export default Statistics;
