+++ src/pages/Profile/ProfilePage.pcss
@@ -1,0 +1,3 @@
+.profile {
+    @apply w-full flex flex-col items-center;
+}
