+++ src/pages/Profile/ProfilePage.tsx
@@ -1,4 +1,8 @@
+import './ProfilePage.pcss';
+
 import { useSelector } from 'react-redux';
+
+import Statistics from '../Statistics/Statistics';
 
 import type { RootState } from '@/store/store';
 
@@ -6,9 +10,9 @@ const ProfilePage = (): JSX.Element => {
   const authState = useSelector((state: RootState) => state.authentication);
 
   return (
-    <section>
-      <h1>ProfilePage</h1>
-      <h2>Hi! {authState.user.name}</h2>
+    <section className="profile">
+      <h2 className="profile_heading">Привет, &nbsp; {authState.user.name}!</h2>
+      <Statistics />
     </section>
   );
 };
